Fix test so it can run with JDK6
